Module Make.Bound

type t =

| Included of Bound_compare.t

| Excluded of Bound_compare.t

| Unbounded

val compare_lower : t -> t -> int

compare_lower bound_a bound_b compares two bounds as lower bounds. Returns an integer less than zero if the first bound is strictly less than the second, zero if the bounds are equal, and an integer greater than zero if the first bound is strictly greater than the second.

val compare_upper : t -> t -> int

compare_upper bound_a bound_b compares two bounds as upper bounds

val min_lower : t -> t -> t

min_lower bound_a bound_b chooses the minimum lower bound between bound_a and bound_b

val max_upper : t -> t -> t

max_upper bound_a bound_b chooses the maximum upper bound between bound_a and bound_b