Module Make.Query_results

type 'a t

val fold : ('acc -> (Interval.t * 'a list) -> 'acc) -> 'acc -> 'a t -> 'acc

fold fn acc results folds over results using function fn, which takes the accumulator and each element as parameters and returns the updated accumulator.

val to_list : 'a t -> (Interval.t * 'a list) list

to_list results transforms results to a list