Parameter Make.1-Bound_compare

type t

val compare : t -> t -> int

compare a b Returns an integer less than zero if the first argument is strictly less than the second, zero if the arguments are equal, and an integer greater than zero if the first argument is strictly greater than the second.